+++ pkg/artifactory/resource/repository/default_repo_layout_map.go
@@ -207,7 +207,7 @@ var defaultRepoLayoutMap = map[string]SupportedRepoClasses{
 		SupportedRepoTypes: map[string]bool{
 			"local":     true,
 			"remote":    true,
-			"virtual":   false,
+			"virtual":   true,
 			"federated": true,
 		},
 	},